Added support for IE.11:
Before this, I used "create-react-app" without having support for I.E.
11. After looking into it, I installed a polyfill for intersection observer
and changed some settings in the config file, but it now works to a
limited extent. The only caveat is that the navbar at the top doesn't
act the same way as it does in Edge, Firefox, or Chrome. Normally in
other browsers, a transparent header is shown at the top to blend in
with the image, and as you slide down an opaque header slides down. See
it in Edge vs I.E. 11 and you'll know what I mean. But I'll settle for
this for now.