Fixed the navbar problem in Internet Explorer:
I used Boolean(document.documentMode) and that must have done the trick,
because the navbar works in I.E. 11 now.

The only problems are that certain images specified in media queries
for the projects section don't load. I tried importing in the
"ThirdPage.js" file as per this link:
https://create-react-app.dev/docs/adding-images-fonts-and-files
It didn't work. This problem shows itself in I.E. 11 and Firefox from
what I see, and things work fine in MS Edge and Chrome.

I believe that this has to do with files being loaded differently in
different browsers. The file structure in firefox when I run "npm start"
is different, and I don't see the images. I could
probably fix this in production.