Changed (hopefully) alignment problem with icons at the bottom:
1) There was a change in the alignment of the icons at the bottom
and that hopefully will show on phones (we'll see). I might have
to make a few commits like this for ths

2) My skills section rates my skills relative to each other. If I've
been coding longer in Python, then I rate it relatively higher compared
to Javascript, which I've only been coding in for a year roughly. So
I changed the heading to reflect this. I also added a disclaimer shown
on devices with size >= 768px that the skills are measured relative to
each other.