Made some refactoring changes:
1) I added a component to './CommonUtilities.js' that is used
throughout the page (the header for each section).

2) I added comments where I deemed necessary for all the JS files.
Some of them don't add much value, some of them do (a balance I'm
still working on).

3) I changed class components to functional components where
I could. I don't know about hooks with REACTjs yet.

4) I removed the code that observes when an accordion section opens
and disables the button while it's opening or closing. It would do
this because when I clicked the button to open or close a section,
you would see changes to the UI, and I didn't want to risk breaking
the UI by double-clicking on the button. It turned out to be
redundant now as I'm observing the "aria-expanded" property and
making UI changes based on this (see the comments in
"SecondPage.js"). It works just fine.

5) I changed the sizing of the Gmail button at the bottom, as it
was too small.

To do:
1) Some more refactoring.
2) Write test cases

However, I've been on this project alone since around February (not
every night, but when I've been coding outside of work, it's only
this), so I'm going to switch to something else for a little while.